update errorMessage component to hook for #107
