+++ packages/app/src/components/index.ts
@@ -6,7 +6,6 @@ export * from './Header';
 export * from './LoadingSpinner';
 export * from './PatientQueue';
 export * from './ProviderFields';
-export * from './ReturnErrorMessage';
 export * from './ScrollToTop';
 export * from './TopAppBar';
 export * from './VideoControls';
